Remove Express API -> Switch to directly netlify functions -> change folder structure -> split dependencies -> change api endpoint name
