+++ .env
@@ -9,4 +9,4 @@ REACT_APP_DEFAULT_DISTRICT=Kernstadt
 REACT_APP_LOCALSTORAGE_KEY=odenwald-garbage-preferences
 REACT_APP_REPOSITORY_URL=$REPOSITORY_URL
 REACT_APP_WEBSITE_URL=https://christian-bechtluft.jimdosite.com/
-REACT_APP_API=$DEPLOY_URL/api
+REACT_APP_API=/.netlify/functions/
