+++ functions/garbages/scrape.js
@@ -1,8 +1,8 @@
 const fetch = require('node-fetch');
 const cheerio = require('cheerio');
 const moment = require('moment');
-const EventDay = require('./models/event-day');
-const allPossibleGarbageBins = require('./constants/garbage-bins');
+const EventDay = require('./eventday');
+const allPossibleGarbageBins = require('./garbagebins');
 
 const baseUrl = 'https://www.abfallkalender-odenwald.de/php/Muellkalender.php?pfad=Deutsch.muellkalender';
